test: Seclude `run` benchmark logic to outer module
